Use int64_t in subtractTime()

+++ src/idletimedetector.cpp
@@ -105,7 +105,7 @@ void IdleTimeDetector::revert(const QDateTime &dialogStart, const QDateTime &idl
 {
     // revert and stop
     QDateTime end = QDateTime::currentDateTime();
-    const int diff = dialogStart.secsTo(end) / secsPerMinute;
+    const int64_t diff = dialogStart.secsTo(end) / secsPerMinute;
     emit subtractTime(idleMinutes + diff); // subtract the time that has been added on the display
     emit stopAllTimers(idleStart);
 }
